+++ lib/analytics.ts
@@ -10,7 +10,6 @@ import {
 var url = require('component-url');
 
 import cloneDeep from 'lodash.clonedeep';
-import pick from 'lodash.pick';
 
 var _analytics = global.analytics;
 
@@ -43,6 +42,7 @@ var nextTick = require('next-tick');
 var normalize = require('./normalize');
 var on = require('component-event').bind;
 var pageDefaults = require('./pageDefaults');
+var pick = require('@ndhoule/pick');
 var prevent = require('@segment/prevent-default');
 var defaults = require('@ndhoule/defaults');
 var store = require('./store');
@@ -599,20 +599,20 @@ Analytics.prototype.page = function(
 
   // Ensure properties has baseline spec properties.
   // TODO: Eventually move these entirely to `options.context.page`
-  const defs = pageDefaults();
+  var defs = pageDefaults();
   defaults(properties, defs);
 
   // Mirror user overrides to `options.context.page` (but exclude custom properties)
   // (Any page defaults get applied in `this.normalize` for consistency.)
   // Weird, yeah--moving special props to `context.page` will fix this in the long term.
-  const overrides = pick(properties, Object.keys(defs));
+  var overrides = pick(Object.keys(defs), properties);
   if (!is.empty(overrides)) {
     options = options || {};
     options.context = options.context || {};
     options.context.page = overrides;
   }
 
-  const msg = this.normalize({
+  var msg = this.normalize({
     properties: properties,
     category: category,
     options: options,