+++ test/analytics.test.ts
@@ -1,4 +1,4 @@
-import assignIn from 'lodash.assignin'
+'use strict';
 
 var Analytics = require('../build').constructor;
 var Facade = require('segmentio-facade');
@@ -6,6 +6,7 @@ var analytics = require('../build');
 var assert = require('proclaim');
 var bind = require('component-event').bind;
 var createIntegration = require('@segment/analytics.js-integration');
+var extend = require('@ndhoule/extend');
 var type = require('component-type');
 var pageDefaults = require('../build/pageDefaults');
 var sinon = require('sinon');
@@ -835,7 +836,7 @@ describe('Analytics', function() {
         assert.deepEqual(opts, { context: { page: defaults } });
         assert.deepEqual(
           props,
-          assignIn({ category: 'category', name: 'name' }, defaults)
+          extend(defaults, { category: 'category', name: 'name' })
         );
         done();
       });